# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AutoGenBook API
  description: HTTP API for AutoGenBook, served by this FastAPI app and reached through
    the nginx-fronted Docker Compose stack. Every route is mounted under /api/v1;
    there is no authentication and no CORS middleware. Error responses use RFC 9457
    problem details (application/problem+json); list endpoints return a {items, total,
    limit, offset} page envelope.
  version: 0.1.0
paths:
...
  /api/v1/files:
    post:
      tags:
      - files
      summary: Upload File
      operationId: files-upload_file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_files-upload_file'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - files
      summary: List Files
      operationId: files-list_files
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: kind
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/FileKind'
          - type: 'null'
          title: Kind
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_File_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/files/{file_id}:
    get:
      tags:
      - files
      summary: Get File
      operationId: files-get_file
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: File Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - files
      summary: Delete File
      operationId: files-delete_file
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: File Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/files/{file_id}/content:
    get:
      tags:
      - files
      summary: Get File Content
      operationId: files-get_file_content
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: File Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects:
    get:
      tags:
      - projects
      summary: List Projects
      operationId: projects-list_projects
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ProjectSummary_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - projects
      summary: Create Project
      operationId: projects-create_project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}:
    get:
      tags:
      - projects
      summary: Get Project
      operationId: projects-get_project
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - projects
      summary: Update Project
      operationId: projects-update_project
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - projects
      summary: Delete Project
      operationId: projects-delete_project
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/duplicate:
    post:
      tags:
      - projects
      summary: Duplicate Project
      operationId: projects-duplicate_project
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/spec:
    get:
      tags:
      - projects
      summary: Get Project Spec
      operationId: projects-get_project_spec
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: format
        in: query
        required: false
        schema:
          type: string
          pattern: ^(txt|json)$
          default: txt
          title: Format
      - name: includeOutline
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Includeoutline
      responses:
        '200':
          description: The project's spec, as plain text (`?format=txt`) or the CLI's
            `book_structure.json`-shaped JSON (`?format=json`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookStructure'
            text/plain:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_files-upload_file:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_files-upload_file
    BookStructure:
      properties:
        title:
          type: string
          title: Title
        summary:
          type: string
          title: Summary
        n_pages:
          type: number
          title: N Pages
        target_readers:
          type: string
          title: Target Readers
        equation_frequency_level:
          type: integer
          title: Equation Frequency Level
        do_consider_outline:
          type: boolean
          title: Do Consider Outline
        do_consider_previous_sections:
          type: boolean
          title: Do Consider Previous Sections
        additional_requirements:
          type: string
          title: Additional Requirements
        max_depth:
          type: integer
          title: Max Depth
        max_output_pages:
          type: number
          title: Max Output Pages
        childs:
          items:
            $ref: '#/components/schemas/BookStructureNode'
          type: array
          title: Childs
      type: object
      required:
      - title
      - summary
      - n_pages
      - target_readers
      - equation_frequency_level
      - do_consider_outline
      - do_consider_previous_sections
      - additional_requirements
      - max_depth
      - max_output_pages
      title: BookStructure
    BookStructureNode:
      properties:
        title:
          type: string
          title: Title
        summary:
          type: string
          title: Summary
        n_pages:
          type: number
          title: N Pages
        needsSubdivision:
          type: boolean
          title: Needssubdivision
        childs:
          items:
            $ref: '#/components/schemas/BookStructureNode'
          type: array
          title: Childs
        structure_locked:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Structure Locked
      type: object
      required:
      - title
      - summary
      - n_pages
      - needsSubdivision
      title: BookStructureNode
...
    File:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        filename:
          type: string
          title: Filename
        contentType:
          type: string
          title: Contenttype
        sizeBytes:
          type: integer
          title: Sizebytes
        sha256:
          type: string
          title: Sha256
        kind:
          $ref: '#/components/schemas/FileKind'
        kbEligible:
          type: boolean
          title: Kbeligible
        createdAt:
          type: string
          format: date-time
          title: Createdat
      type: object
      required:
      - id
      - filename
      - contentType
      - sizeBytes
      - sha256
      - kind
      - kbEligible
      - createdAt
      title: File
    FileKind:
      type: string
      enum:
      - upload
      - artifact
      title: FileKind
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OutlineTreeReplaceNode:
      properties:
        title:
          type: string
          minLength: 1
          title: Title
        summary:
          type: string
          title: Summary
          default: ''
        targetPages:
          anyOf:
          - type: number
          - type: 'null'
          title: Targetpages
        subPrompt:
          anyOf:
          - type: string
          - type: 'null'
          title: Subprompt
        mathLevel:
          $ref: '#/components/schemas/MathLevel'
          default: rigorous
        equationDensityLevel:
          anyOf:
          - type: integer
            maximum: 5.0
            minimum: 1.0
          - type: 'null'
          title: Equationdensitylevel
        children:
          items:
            $ref: '#/components/schemas/OutlineTreeReplaceNode'
          type: array
          title: Children
      additionalProperties: false
      type: object
      required:
      - title
      title: OutlineTreeReplaceNode
      description: 'One node of a full outline replacement, nested, without ids (ids
        are

        assigned on write).'
    OutputFormat:
      type: string
      enum:
      - markdown
      - latex
      - pdf
      title: OutputFormat
    Page_File_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/File'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: Page[File]
...
    Page_ProjectSummary_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ProjectSummary'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: Page[ProjectSummary]
...
    Project:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        title:
          type: string
          title: Title
        subtitle:
          type: string
          title: Subtitle
        authors:
          items:
            type: string
          type: array
          title: Authors
        topic:
          type: string
          title: Topic
        targetAudience:
          $ref: '#/components/schemas/TargetAudience'
        totalPagesBudget:
          type: integer
          title: Totalpagesbudget
        equationFrequencyLevel:
          type: integer
          title: Equationfrequencylevel
        doConsiderOutline:
          type: boolean
          title: Doconsideroutline
        doConsiderPreviousSections:
          type: boolean
          title: Doconsiderprevioussections
        outputFormat:
          $ref: '#/components/schemas/OutputFormat'
        maxOutlineLevels:
          type: integer
          title: Maxoutlinelevels
        additionalRequirements:
          anyOf:
          - type: string
          - type: 'null'
          title: Additionalrequirements
        sources:
          items:
            $ref: '#/components/schemas/Source'
          type: array
          title: Sources
        outline:
          items:
            $ref: '#/components/schemas/OutlineNodeTree'
          type: array
          title: Outline
        lastRunId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Lastrunid
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
      type: object
      required:
      - id
      - title
      - subtitle
      - authors
      - topic
      - targetAudience
      - totalPagesBudget
      - equationFrequencyLevel
      - doConsiderOutline
      - doConsiderPreviousSections
      - outputFormat
      - maxOutlineLevels
      - createdAt
      - updatedAt
      title: Project
    ProjectCreate:
      properties:
        title:
          type: string
          minLength: 1
          title: Title
        subtitle:
          type: string
          title: Subtitle
        authors:
          items:
            type: string
            minLength: 1
          type: array
          title: Authors
        topic:
          type: string
          minLength: 1
          title: Topic
        targetAudience:
          $ref: '#/components/schemas/TargetAudience'
          default: graduate
        totalPagesBudget:
          type: integer
          maximum: 2000.0
          minimum: 5.0
          title: Totalpagesbudget
          default: 350
        equationFrequencyLevel:
          type: integer
          maximum: 5.0
          minimum: 1.0
          title: Equationfrequencylevel
          default: 4
        doConsiderOutline:
          type: boolean
          title: Doconsideroutline
          default: true
        doConsiderPreviousSections:
          type: boolean
          title: Doconsiderprevioussections
          default: true
        outputFormat:
          $ref: '#/components/schemas/OutputFormat'
          default: markdown
        maxOutlineLevels:
          type: integer
          maximum: 5.0
          minimum: 1.0
          title: Maxoutlinelevels
          default: 3
        additionalRequirements:
          anyOf:
          - type: string
          - type: 'null'
          title: Additionalrequirements
        sources:
          anyOf:
          - items:
              $ref: '#/components/schemas/SourceCreate'
            type: array
          - type: 'null'
          title: Sources
        outline:
          anyOf:
          - items:
              $ref: '#/components/schemas/OutlineTreeReplaceNode'
            type: array
          - type: 'null'
          title: Outline
      type: object
      required:
      - title
      - subtitle
      - authors
      - topic
      title: ProjectCreate
    ProjectSummary:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        title:
          type: string
          title: Title
        subtitle:
          type: string
          title: Subtitle
        authors:
          items:
            type: string
          type: array
          title: Authors
        topic:
          type: string
          title: Topic
        targetAudience:
          $ref: '#/components/schemas/TargetAudience'
        totalPagesBudget:
          type: integer
          title: Totalpagesbudget
        sourcesCount:
          type: integer
          title: Sourcescount
        outlineNodeCount:
          type: integer
          title: Outlinenodecount
        lastRunId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Lastrunid
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
      type: object
      required:
      - id
      - title
      - subtitle
      - authors
      - topic
      - targetAudience
      - totalPagesBudget
      - sourcesCount
      - outlineNodeCount
      - createdAt
      - updatedAt
      title: ProjectSummary
    ProjectUpdate:
      properties:
        title:
          anyOf:
          - type: string
            minLength: 1
          - type: 'null'
          title: Title
        subtitle:
          anyOf:
          - type: string
          - type: 'null'
          title: Subtitle
        authors:
          anyOf:
          - items:
              type: string
              minLength: 1
            type: array
          - type: 'null'
          title: Authors
        topic:
          anyOf:
          - type: string
            minLength: 1
          - type: 'null'
          title: Topic
        targetAudience:
          anyOf:
          - $ref: '#/components/schemas/TargetAudience'
          - type: 'null'
        totalPagesBudget:
          anyOf:
          - type: integer
            maximum: 2000.0
            minimum: 5.0
          - type: 'null'
          title: Totalpagesbudget
        equationFrequencyLevel:
          anyOf:
          - type: integer
            maximum: 5.0
            minimum: 1.0
          - type: 'null'
          title: Equationfrequencylevel
        doConsiderOutline:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Doconsideroutline
        doConsiderPreviousSections:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Doconsiderprevioussections
        outputFormat:
          anyOf:
          - $ref: '#/components/schemas/OutputFormat'
          - type: 'null'
        maxOutlineLevels:
          anyOf:
          - type: integer
            maximum: 5.0
            minimum: 1.0
          - type: 'null'
          title: Maxoutlinelevels
        additionalRequirements:
          anyOf:
          - type: string
          - type: 'null'
          title: Additionalrequirements
      additionalProperties: false
      type: object
      title: ProjectUpdate
...
    SourceCreate:
      properties:
        fileId:
          type: string
          format: uuid
          title: Fileid
        type:
          anyOf:
          - $ref: '#/components/schemas/SourceType'
          - type: 'null'
        authors:
          anyOf:
          - type: string
          - type: 'null'
          title: Authors
        year:
          anyOf:
          - type: string
            pattern: ^\d{4}$
          - type: 'null'
          title: Year
        doi:
          anyOf:
          - type: string
          - type: 'null'
          title: Doi
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - fileId
      title: SourceCreate
...
    TargetAudience:
      type: string
      enum:
      - undergraduate
      - graduate
      - phd_researcher
      - industry_practitioner
      title: TargetAudience
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError